+++ internal/pkg/dbmigrate/migrations/000015_datasource_version.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE datasources DROP COLUMN version;