+++ internal/pkg/dbmigrate/migrations/000015_datasource_version.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE datasources ADD COLUMN version VARCHAR(128) DEFAULT '' AFTER is_enabled;